+++ src/engine/utility/jconv.cpp
@@ -116,12 +116,12 @@ namespace jconv {
 
 	    str.replace(pos, from_.size(), to_);
 
-	    return this->operator()(str, remain - from_.size(), pos + to_.size());
+	    return this->operator()(str, (unsigned)(remain - from_.size()), (unsigned)(pos + to_.size()));
 	}
     };
 
     static void kana_convert(kana_member target, kana_member replacement, std::string& str) {
-	unsigned remain = str.size();
+	unsigned remain = (int)str.size();
 
 	for(kana** tbl = kana_tables; *tbl != 0; ++ tbl) {
 	    for(kana* ptr = *tbl; remain && ptr->hirakana; ++ ptr) {
